[apps] Refactor exp, cos, log
